% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_files.R
\name{copy_testcase}
\alias{copy_testcase}
\title{copy template files to a specified folder}
\usage{
copy_testcase(folder = ".")
}
\arguments{
\item{folder}{the folder where the test case files should be copied to}
}
\description{
This function copies the files for the test case of Lake Bryrup, Denmark provided on the gitlab page of WET
}
\examples{
\dontrun{
copy_testcase(".")
}

}
\author{
Johannes feldbauer
}
\keyword{methods}
