% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wet.R
\name{run_wet}
\alias{run_wet}
\title{Run the GOTM-WET model}
\usage{
run_wet(
  sim_folder = ".",
  yaml_file = "gotm.yaml",
  verbose = FALSE,
  args = character()
)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{yaml_file}{filepath; to file with GOTM-WET setup. Defaults to
'gotm.yaml'}

\item{verbose}{Save output as character vector. Defaults to FALSE}

\item{args}{Optional arguments to pass to GOTM-WET executable}
}
\description{
This runs the GOTM-WET model on the specific simulation stored in
\code{sim_folder}. The specified \code{sim_folder} must contain
valid NML files.
}
\examples{
copy_testcase(".")
run_wet()
}
\author{
Tadhg Moore, Johannes Feldbauer
}
\keyword{methods}
