% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_version.R
\name{wet_version}
\alias{wet_version}
\title{Return the current GOTM-WET model version}
\usage{
wet_version()
}
\description{
Returns the current version of the GOTM-WET model being used
}
\examples{
print(wet_version())


}
\author{
Tadhg Moore, Johannes Feldbauer
}
\keyword{methods}
